+++ tests/WildcardTest.php
@@ -1,0 +1,43 @@
+<?php
+
+namespace hamburgscleanest\GuzzleAdvancedThrottle\Tests;
+
+use hamburgscleanest\GuzzleAdvancedThrottle\Wildcard;
+use PHPUnit\Framework\TestCase;
+
+/**
+ * Class WildcardTest
+ * @package hamburgscleanest\GuzzleAdvancedThrottle\Tests
+ */
+class WildcardTest extends TestCase
+{
+
+    /** @test
+     */
+    public function matches_wildcards() : void
+    {
+        $this->assertTrue(Wildcard::matches('test1.{wildcard1}.test2.{wildcard2}.test3', 'test1.replace1.test2.replace2.test3'));
+    }
+
+    /** @test
+     */
+    public function does_not_match_when_not_all_wildcards_given() : void
+    {
+        $wildcardText = 'test1.{wildcard1}.test2.{wildcard2}.test3';
+
+        $this->assertFalse(Wildcard::matches($wildcardText, 'no_wildcards'));
+        $this->assertFalse(Wildcard::matches($wildcardText, 'test1.{wildcard1}.test2.test3'));
+        $this->assertFalse(Wildcard::matches($wildcardText, 'test1.{wildcard1}.test2.{wildcard2}'));
+        $this->assertFalse(Wildcard::matches($wildcardText, 'one.{wildcard1}.two.{wildcard2}.three'));
+    }
+
+    /** @test
+     */
+    public function ignores_text_without_wildcard() : void
+    {
+        $wildcardText = 'test1';
+
+        $this->assertTrue(Wildcard::matches($wildcardText, $wildcardText));
+        $this->assertFalse(Wildcard::matches($wildcardText, 'test2'));
+    }
+}
